Organize JS better
