+++ gulpfile.babel.js
@@ -6,9 +6,9 @@ import rollupConfig from './rollup.config';
 
 gulp.task('run_rollup', () => {
   return gulp.src([
-    'src/core.js',
-    'src/compatibility.js',
-    'panels/automation/editor.js',
+    'js/core.js',
+    'js/compatibility.js',
+    'js/editor/editor.js',
     'demo_data/demo_data.js',
   ])
   .pipe(rollupEach(rollupConfig))
@@ -25,8 +25,7 @@ gulp.task('ru_all', ['run_rollup'], () => {
 
 gulp.task('watch_ru_all', ['ru_all'], () => {
   gulp.watch([
-    'src/**/*.js',
-    'panels/**/*.js',
+    'js/**/*.js',
     'demo_data/**/*.js'
   ], ['ru_all']);
 });
