+++ js/editor/script/index.js
