+++ js/editor/trigger/index.js
