+++ js/editor/trigger/numeric_state.js
